Fix watch flag in commands other than 'build' (#153)
With the recent changes we added more "build" actions inside other commands,
which implies also accepting --watch in these other commands (expected behaviour: 
on file change rerun the command that is being watched).

However, the watch was only piped to the build command, so only the build part would 
be rerun on change. This fixes that by running the whole action on file change instead.