Add spago-curator (#202)
This adds a `spago-curator` executable that we can use to automate some CI.

The first command for which this is useful is `spago-curator index-github-meta`

When you run it (provided that you have a GitHub token in your env), `spago-curator`
will take the latest package set and crawl GitHub downloading the list of all tags and
commits for every repo in the set, and put the result in a `metadataV1.json` file.

The idea is that @spacchettibotti will run this command periodically and upload the 
resulting file to the `package-sets-metadata` repo, so that in `spago` we can:
* download this file if missing from the global cache or older than let's say 1 day
* then whenever we have to download a package we can check if the tag or commit
  is in the cached GitHub index, and if it is then we can "globally cache" the package.

A question that might arise is: "why not just hit GitHub asking for this info instead of 
having this complicated CI process?"

The problem is that GitHub limits token-less API requests to 50 per hour, so any
decently-sized installation will fail to get all the "cacheable" items, making the global
cache kind of useless. So we are just caching all of that info for everyone here.

The other solution would be to require people to have a GitHub token in the env, but
that's not nice for UX nor secure.