Add global packages cache and allow commit hashes (#188)
It works so that we download a file containing metadata about commits and tags
of the packages in the package set, and see if the refs we want to download are there:
https://github.com/spacchetti/package-sets-metadata/blob/master/metadataV1.json

If they are we also copy them to this global cache once they download, so we can reuse
them in other projects.

An additional optimization that this introduces is that since we are only downloading from
GitHub we can point to repo archives, so avoiding to use the `git` commands for everything 
(which are really slow). 
This means that we can also fetch archives with commit refs.
Now, if a package is not "globally cacheable" the old git method will be used in the local
cache, and here we make it slightly slower (because we `git fetch` all the refs instead of
just one), but allow to use commit hashes as refs too.

Another thing that we fix here is the "temporary directory" problem: before this we were
fetching things inside the local cache directory, possibly giving "broken downloads" if the
command failed. Now we do everything in temp directories, which makes is slightly more
complicated but makes sure the result is always correct.