Fix psc-package-clean and test cleanup on Windows
Replaces use of `Turtle.rmtree` with `System.Directory.removePathForcibly`
to fix the following issues on Windows.

`spago psc-package-clean` (and its test) fail because:
1. Git marks files under its /objects subdirectory as read-only
2. Turtle.rmtree fails to delete files marked as read-only

So `rmtree` fails to remove any folder that's a git repository.

As well as `psc-package-clean`, general test cleanup fails, meaning the
test suite will fail completely when run for the second time.