Read purs version from inside the set instead of its GitHub tag (#253)
Before this `spago` would try to find out which `purs` version is supported by
the set by reading the GitHub tag the upstream comes from. This brings a whole 
bunch of issues, e.g. we cannot ensure anything if the package set comes from
somewhere else while the check is hard to override if we do, etc.

With this change we now read the "minimum supported purs version" just as the
version of one of the packages in the set.

This fixes #225, because we:
- fix the error message to just say "SemVer mismatch" instead of newer/older
- make the "minimum purs version" easily overridable, as it's just the version of
  a package, so there's no need for a flag
- make this version check actually robust, as we're not trying to climb the Dhall
  import chain from the AST anymore