Handle branch names differing only by case
Escape upper case characters in branch names to ensure that branches
which only differ in case don't cause issues on case-insensitive file
systems.

Also switch to using an allowlist for escaping characters: it seems a
little safer to specify the characters which don't need escaping (and to
escape everything else), than to specify characters which do need
escaping.

Finally, add a test that `getCacheVersionDir` is injective.