Fix watch paths on Windows (#387)
`Watch.globToParent` was producing paths on Windows that lacked a slash
after the drive letter, ie: `C:Path\To\Folder`. These are treated as
relative paths on Windows, causing invalid paths as described in
https://github.com/spacchetti/spago/issues/380#issuecomment-524583489

Using `System.FilePath.splitDirectories` instead to split the path ensures that
the path is properly split.