Disable output sharing (#526)
This commit disables output sharing that was introduced in #422. 

The reasoning for this is twofold:
1. this change seems to break `psc-ide` and various other tooling
2. Spago messing with people's `output` folder doesn't really feel right,
  because e.g. it forces people to depend on wherever Spago decides the
  output folder to be in their project

We could of course just invert the logic of the flag and make it opt-in
instead, but I just don't think we should encourage point (2) above.

The "proper" way to do this IMHO is that Spago could do some kind of
"build artifact caching": i.e. sharing compiled JS between projects
compiled with the same set of packages and compiler.
This is a strictly better solution, because:
- it doesn't mess with people's output folder
- it enables artifact sharing even between different projects and not
  just in the same monorepo