Send watch-mode messages to the controlling terminal.
- Send watch-mode prompts directly to the controlling terminal
  (on non-Windows) rather than to stderr.

- Make stderr, not stdout, the fallback terminal handle.

- Check "TERM" environment var to set color usage in watch mode.

Motivation for change:

While Spago is running in watch mode, messages sent to the terminal act
like a simple interface for user interaction. By listing options and
pointing out issues and events, these messages aren't typical logging
diagnostics but instead have an important role in Spago's behavior as an
application. However, although they're essential to the proper
functioning of watch mode, and although, ordinarily, for this reason,
they wouldn't be sent to stderr but to stdout, Spago's interaction with
the 'purs' compiler makes using stdout problematic. Specifically, 'purs'
also sends data to stdout, and parsing that data is overly cumbersome
when watch-mode prompts are mixed in too. Even so, continuing to send
the prompts to stderr has its own drawbacks: first, it runs counter to
the conventional purpose of stderr, and, worse, the --quiet flag,
a CLI option that users should expect to be available, suppresses watch
mode's messages, making watch mode confusing and nearly nonfunctional.

A possible solution to these complications, and the one chosen here,
is to skip the standard terminal handles except for diagnostic logging
and to have watch mode emit prompts to the controlling terminal
directly. Although this may not be a conventional approach for command-
line programs, the multi-channel I/O model
    { stdout   -> data for both purescript and spago
    , stderr   -> diagnostics for both purescript and spago
    , terminal -> UI/UX for spago's watch mode
    }
facilitates data usage without sacrificing the availability of watch
mode's terminal-based user interface.