Better environments (#610)
This (rather large, sorry for that) refactoring completes the work 
of introducing RIO that started in #486 

The reason for this PR is to tidy up the "execution requirements"
of every command - i.e. what does every command need in order to run.
For how the code is structured now this will allow us to request
"capabilities" in functions where we need them.
Example: there's now a `HasPurs` typeclass that gives the path of the
compiler. If there's no `purs` we now fail right at the beginning with a
decent error message, rather than deep in the execution tree when we
actually need to run it.

This is a breaking change because it removes the deprecated
`--no-share-output` flag for build commands.