Fix finding psa on Windows (#693)
In `getPurs` we were just using the standard `Directory.findExecutable`
to look for `psa`, forgetting to check for `psa.cmd` too. In windows
this was looking for `psa.exe`, which doesn't exist.

To help prevent this happening again, a self-written `findExecutable`
function in `Prelude` is added, instead of just exporting
`Directory.findExecutable`. This new version will always first check
for a `.cmd` version of the executable name on Windows.