Fix colors on Windows (#768)
Fixes colors on Windows (#749), by calling the `hSupportsANSIWithoutEmulation` function, which enables terminal colors on Windows 10+ (https://hackage.haskell.org/package/ansi-terminal-0.11/docs/System-Console-ANSI.html#v:hSupportsANSIWithoutEmulation)

Additionally, using this function in the way we do here means colors are automatically disabled when output is redirected to a file. Test fixtures are modified to reflect this.

`hSupportsANSIWithoutEmulation` should also check for the `TERM=dumb` environment variable, but it doesn't on Windows, and on *nix this hasn't been tested, so the existing code here that does this check has been left in for now.

Finally, support is added for the https://no-color.org spec with a simple check for a `NO_COLOR` environment variable.