Disable psa for some tests
Some tests fail if the system has `psa` installed, because it affects
the outputs of commands run. The difference is immaterial; namely just
the addition of a table of number of warnings/errors at the end of
output.

Explicitly disables psa for these tests with the `--no-psa` flag, so
they will also pass if it is installed.