Implement let-binding AST handling
In this commit, I moved the code that checks whether
the user-requested packges exist in the package set
into the `withRawConfigAST`, so that the new
`Spago.Config.AST` module doesn't introduce a cyclial dependency with
the `Spago.Config` module.

I also provide both the raw and normalized AST expressions
as it's the only way to correctly determine which packages
haven't yet been installed.