Use semantic equivalent check for config values
Config's publishedConfig value will store a
Left (RequiredKeyMissing k kvs) when either the
license or repository key does not successfully parse.
This itself is not an issue. However, the `kvs` value
is a Dhall.Map.Map value, which does honor insertion order.
Thus, two maps that contain the same thing but store them in two
slightly different orders are considered unequal.
So, we sort those values before we check whether they are equal.