Enable WithinField to track let bindings
If we are updating the dependencies field,
and have multiple let bindings
(e.g. dependencies = let x = ... let y = x let z = y in z),
then we should update the x binding's value.
If that value refers to a binding outside of the
scope of the dependencies key, then we stop
and do the update within x binding value.
Otherwise, we might introduce a side-effect in case
the x binding's value is used elsewhere in the expression.

Normally, we would use an Int to track the
scope depth of the let bindings.
In other words, `x` above would be 0, `y` would be `1` and
`z` would be `2`. When we reach `z`, we go back to `y` because `z`'s scope depth is not 0.
When we reach `y`, we got back to `x` becaues `y`'s scope depth is not 0.
Because `x'`s scope depth is 0, we stop and do the update.

However, if we have an expression like
dependencies = let x = { deps = [] } in x.deps
then we also need to keep track of a new key stack
that we only discover once we are WithinField.
Thus, rather than using `0`, we'll use `[Text]`.
An empty list corresponds to the 0 scope depth,
and a non-empty list means we can traverse back up
the expression if needed