VariableName second arg: `NonEmpty Text` -> `[Text]`
This change enables us to pass the keystack
back up to the let binding expression that defines
the value for the binding. However, since the
key stack might be empty, we can no longer use
`NonEmpty Text`. Unfortunately, this change
means `SearchingforField`'s arg must also be updated
and by implication, `WithinField` and `SearchingForField` are
conceptually the same thing.