Optimize code: convert ExprLevel to a keyStack
To maintain the previous behavior, a let binding
should only allow a `VariableName` to propagate to its parent expression
if its own level is not an empty key stack (which means
we're at the field we need to update). Otherwise,
we will traverse outside of the current expression,
and introduce a side-effect if we update the
actual let binding's value.