Remove redundant first case for With
Previously, we were checking whether we should only
update the `updateExpr` and only then falling back to updating
either the `updateExpr` or the `recordExpr`.
If the `updateExpr` was an exact match of the remaining
fields we were seaching for, we would go down that path.
Otherwise (i.e. the `updateExpr` was a partial match), we would try
updating the `updateExpr` and then the `recordExpr` if the former failed
By removing the "full" updateExpr case, we get the same execution but without
the redundant code