Fix 2+ sequential config ast updates by changing eq check
When 2+ config AST updates occur, the second one does not have access
to the result of the first config update. Thus,
the `isSemanticallyEquivalent` check fails because
the second config update's expectec config
does not contain the updates done in the first one.
For example, if we update the name and then add dependencies,
the dependencies ast update check does not have
the new name in the 'expectedConfig' value.

Moreover, in the process of doing this change,
I discovered that I could remove the 'sortDependencies' part of the
publishConfig check by just verifying that the
keys in each maps are the same since that's all
the error really cards about: missing keys.