Refactor Fetch.run to support Build.run's multi-package build
- Fetch.run returns back the same value as it
did before (i.e. `dependencies`) but now under
the `dependencies` field name.
Under the `packageDependencies` field name,
it returns a map that can be used to look up
all the dependencies for a given workspace package.
This is used in Build.run to get all the dependencies
for a given package. However, to produce the `packageDependencies` value,
we have to call `getTransitiveDeps` multiple times
and don't get to share the work done in previous calls.
So, if 5 packages all depend on `halogen`, we're recomputing
`halogen`'s transitive dependencies 5 times.

- Fetch.run's internal implementation was refactored a bit
to prevent the impossible case where `workspace.selected` is `Just`
but calling what was `getConfigPath` was `Nothing`.