Fix output truncation with json-errors and build failure. Fix #1160
Problematic situation is when --json-errors is passed (output is one very long line printed out all at once),
there are many warnings (json > 65536 bytes) and there is an error (spago should exit with an error code).
Set the error code and allow control flow to naturally exit to allow output to be flushed.