Make "missing packages" error message more descriptive
Currently there exists confusion due to Pursuit not showing package
sets that include a package. IOW it always looks as if every package
on Pursuit available on every "package set" from the registry. As
result, when a user receives error message from Spago that says that a
package wasn't found, unless a user already knows this situation, they
unlikely would think it's a problem on the remote side rather than
with syntax or something similar.

But on the `spago` side we can reduce confusion by showing a user
a message hinting that it is possible to enlist available packages
with `spago ls packages`.

Fixes: https://github.com/purescript/spago/issues/1287